Fix and update tests

+++ app/routes/admin.js
@@ -14,22 +14,34 @@ const utils = require(process.env.PWD + '/app/lib/utils');
  */
 exports.post = async (request, response) => {
   const getResult = async request => {
-    // Determine admin action, ensuring token is verified and provides scope
-    const accessToken = request.headers.authorization;
-    const authResponse = await indieauth.verifyToken(accessToken);
-    const {scope} = authResponse;
     const {query} = request;
 
-    if (!scope) {
+    // Ensure request includes query (and for now, that for purging cache)
+    const hasQuery = Object.entries(query).length !== 0;
+    if (!hasQuery || query.purge !== 'cache') {
+      return utils.error('not_found');
+    }
+
+    // Verify access token
+    const accessToken = request.headers.authorization;
+    const authResponse = await indieauth.verifyToken(accessToken);
+    const authError = authResponse.body && authResponse.body.error;
+
+    // Return any errors from IndieAuth token endpoint
+    if (authError) {
       return authResponse;
     }
 
-    if (scope.includes('delete') && query.purge === 'cache') {
+    // Ensure token provides enough scope
+    const {scope} = authResponse;
+    const hasScope = scope && scope.length > 0;
+
+    if (hasScope && (scope.includes('delete') && query.purge === 'cache')) {
       cache.delete();
       return utils.success();
     }
 
-    return utils.error('not_found');
+    return utils.error('insufficient_scope');
   };
 
   try {
